+++ test/test_get_inventory.py
@@ -20,6 +20,9 @@ def test_get_inventory():
         TEST_DIR
         / "platform/cisco_nxos/show_inventory_pipe_json_pretty/show_inventory_pipe_json_pretty.txt"
     ).read_text(encoding="utf-8")
+    cisco_ios_data = (
+        TEST_DIR / "platform/cisco_ios/show_inventory/show_inventory.txt"
+    ).read_text(encoding="utf-8")
     a10_data = (TEST_DIR / "platform/a10/show_hardware/show_hardware.txt").read_text(
         encoding="utf-8"
     )
@@ -33,6 +36,7 @@ def test_get_inventory():
         k in template_inputs for k in [
             "arista_eos_inventory",
             "a10_show_hardware",
+            "cisco_ios_inventory",
             "cisco_nxos_inventory",
             "cisco_xr_inventory",
             "juniper_junos_inventory",
@@ -40,6 +44,7 @@ def test_get_inventory():
     ), "Template inputs are wrong"
 
     parser.add_input(template_name="cisco_xr_inventory", data=cisco_xr_data)
+    parser.add_input(template_name="cisco_ios_inventory", data=cisco_ios_data)
     parser.add_input(template_name="cisco_nxos_inventory", data=cisco_nxos_data)
     parser.add_input(template_name="a10_show_hardware", data=a10_data)
     parser.parse()
@@ -76,6 +81,32 @@ def test_get_inventory():
             "slot": "Power Supply 2",
         },
     ], "cisco_nxos_inventory parsing results are wrong"
+    assert result["cisco_ios_inventory"] == [
+        {
+            "description": "3640 chassis",
+            "module": "",
+            "serial": "FF1045C5",
+            "slot": "3640 chassis",
+        },
+        {
+            "description": "One port Fastethernet TX",
+            "module": "NM-1FE-TX=",
+            "serial": "7720321",
+            "slot": "One port Fastethernet TX",
+        },
+        {
+            "description": "One port Fastethernet TX",
+            "module": "NM-1FE-TX=",
+            "serial": "7720321",
+            "slot": "One port Fastethernet TX",
+        },
+        {
+            "description": "One port Fastethernet TX",
+            "module": "NM-1FE-TX=",
+            "serial": "7720321",
+            "slot": "One port Fastethernet TX",
+        },
+    ], "cisco_ios_inventory parsing results are wrong"
     assert result["a10_show_hardware"] == [
         {
             "description": "Thunder Series Unified Application Service Gateway",